requirements.txt
lxml
pillow
PyQt5
PyQt5-Qt5
PyQt5_sip
python-docx
pywin32